FUCK THIS SHIT IM BORED also problem 112 added

+++ src/problems/p1.hpp
@@ -1,20 +1,21 @@
 #pragma once
 
 //std
+#include <iostream>
 #include <unordered_map>
 #include <vector>
 
 //lib
 #include "util.hpp"
 
-namespace p1 {
+ namespace p1 {
 	class Solution {
 	public:
 		std::vector<int> twoSum(const std::vector<int>& nums, int target) const {
 			std::vector <int> result(2);
 			std::unordered_map<int, int> c;
 
-			for (size_t i = 0; i != nums.size(); ++i) {
+			for (int i = 0; i != nums.size(); ++i) {
 				auto it = c.find(target - nums[i]);
 				if (it != c.end() && it->second != i) {
 					result[0] = it->second < i ? it->second : i;
